Build score function from heuristic

(Internal) Takes a heuristic object and turns it into a function that takes two arguments, entropy and number of blocks that is used to calculate the score for a given entropy-number of blocks step for agglomerative merging algorithms.

build_score_fn(heuristic)

Arguments

heuristic
How the best partitioning is defined. Takes either a function that takes one/two arguments: an entropy vector and an optional number of blocks vector with each element corresponding to a given location, or a string labeling algorithm. Currently only "lowest", "dev_from_rolling_mean", and "nls_residual" are supported.

Value

A function that takes and entropy and number of block vector and returns a score for partitioning (higher = better)

Examples

# Setup fake entropy and number of blocks vectors
entropy <- -(10:1)*1000 + rnorm(10, 200)
num_blocks <- 1:10

# Works with heuristic functions that take two arguments
nls_score <- function(e, k){
  entropy_model <- nls(e ~ a + b * log(k), start = list(a = max(e), b = -25))
  -residuals(entropy_model)
}

build_score_fn(nls_score)(entropy, num_blocks)

#>  [1] -1436.7564   286.4281   883.0498  1011.5289   888.1043   604.3041
#>  [7]   213.2580  -263.3895  -800.3742 -1386.1516
#> attr(,"label")
#> [1] "Residuals"

# Works with functions that take one argument
invert_score <- function(e) -e/2
build_score_fn(invert_score)(entropy, num_blocks)

#>  [1] 4900.7000 4399.8723 3901.2186 3400.0028 2899.6892 2399.4258 1900.9109
#>  [8] 1400.1237  900.1221  400.1414

# Works with predefined strings
build_score_fn("dev_from_rolling_mean")(entropy, num_blocks)

#>  [1]     0.0000  -500.8277  -998.7567 -1500.8913 -1501.0130 -1501.3166
#>  [7] -1498.1925 -1499.8275 -1500.0470 -1500.3663

build_score_fn("lowest")(entropy, num_blocks)

#>  [1] 9801.4000 8799.7447 7802.4373 6800.0056 5799.3784 4798.8516 3801.8218
#>  [8] 2800.2473 1800.2442  800.2827