Choose and load best model state from agglomerative collapsing algorithm

Choose and load best model state from agglomerative collapsing algorithm

choose_best_collapse_state(
  sbm,
  collapse_results,
  heuristic = "dev_from_rolling_mean",
  verbose = FALSE
)

Arguments

sbm
SBM model object as created by create_sbm().

collapse_results
Results of running agglomerative collapse algorithm on sbm with collapse_groups.

heuristic
How the best partitioning is defined. Takes either a function that takes one/two arguments: an entropy vector and an optional number of groups vector with each element corresponding to a given location, or a string labeling algorithm. Currently only lowest and dev_from_rolling_mean are supported.

verbose
Should model tell you what step was chosen (TRUE or FALSE)?

Value

SBM model updated to match state from the merge step with 'best' partitioning.

Examples

set.seed(42)

# Start with a random network of two groups with 25 nodes each
network <- sim_basic_block_network(n_groups = 3, n_nodes_per_group = 25)

# Create SBM from simulated data
my_sbm <- create_sbm(network)

# Run agglomerative clustering with no intermediate MCMC steps on network
collapse_results <- collapse_run(my_sbm, sigma = 3, start_group_num = 1, end_group_num = 6)

# Choose best result with default heuristic
my_sbm <- choose_best_collapse_state(my_sbm, collapse_results, verbose = TRUE)

#> Choosing collapse with 3 groups and an entropy of 2032.81284279538.

# Score heuristic that fits a nonlinear model to observed values and chooses by largest negative residual
nls_score <- function(e, k){
  entropy_model <- nls(e ~ a + b * log(k), start = list(a = max(e), b = -25))
  -residuals(entropy_model)
}

# Choose result using custom heuristic function
my_sbm <- choose_best_collapse_state(my_sbm,
                                     collapse_results,
                                     heuristic = nls_score,
                                     verbose = TRUE)

#> Choosing collapse with 3 groups and an entropy of 2032.81284279538.