Get collapse results from model

Retrieves the dataframe recording the results of running either collapse_blocks or collapse_run on sbm_network object.

get_collapse_results(sbm)

Arguments

sbm
Object of class sbm_network.

Value

List with two dataframes. The first telling for all sweeps everytime a node was moved and what group it was moved to. The second telling for each sweep the entropy delta and total number of nodes that were moved to new groups in that sweep. If track_pairs = TRUE, then an additional pairing_counts dataframe is added to output.

Examples

# Start with a random network of two blocks with 25 nodes each and
# run agglomerative clustering with no intermediate MCMC steps on network
my_sbm <- sim_basic_block_network(n_blocks = 2, n_nodes_per_block = 25) %>%
  collapse_blocks(num_mcmc_sweeps = 0, sigma = 1.1)

# Look at the results of the collapse directly
get_collapse_results(my_sbm)

#>    n_blocks entropy_delta  entropy                        merges
#> 1        45     19.197604 211.2916 1, 3, 2, 4, 5, 6, 9, 10, 7, 8
#> 2        40     41.231868 252.5234 1, 2, 5, 4, 3, 9, 7, 6, 8, 10
#> 3        36     32.903798 285.4272        4, 3, 2, 1, 6, 7, 8, 5
#> 4        32     26.814067 312.2413        4, 3, 1, 2, 6, 7, 8, 5
#> 5        29     28.481612 340.7229              2, 1, 3, 4, 6, 5
#> 6        26     18.959531 359.6824              2, 1, 3, 6, 5, 4
#> 7        23     16.215916 375.8984              1, 3, 2, 5, 4, 6
#> 8        20     17.771403 393.6698              3, 1, 2, 6, 5, 4
#> 9        18      8.248919 401.9187                    1, 2, 4, 3
#> 10       16      8.484275 410.4029                    2, 1, 3, 4
#> 11       14     12.413291 422.8162                    1, 2, 4, 3
#> 12       12     13.358533 436.1748                    1, 2, 3, 4
#> 13       10     12.385142 448.5599                    2, 1, 3, 4
#> 14        9      5.632665 454.1926                          1, 2
#> 15        8      5.308673 459.5013                          1, 2
#> 16        7      7.265302 466.7666                          1, 2
#> 17        6      6.316257 473.0828                          1, 2
#> 18        5      7.805147 480.8880                          1, 2
#> 19        4      7.073156 487.9611                          1, 2
#> 20        3     10.415626 498.3767                          1, 2
#> 21        2      3.603375 501.9801                          1, 2
#> 22        1    178.079000 680.0591                          1, 2
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        state
#> 1  g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_0, bl_node_1, bl_node_2, bl_node_3, bl_node_4, bl_node_5, bl_node_6, bl_node_7, bl_node_8, bl_node_9, bl_node_10, bl_node_11, bl_node_12, bl_node_13, bl_node_14, bl_node_15, bl_node_16, bl_node_17, bl_node_13, bl_node_19, bl_node_20, bl_node_7, bl_node_22, bl_node_23, bl_node_24, bl_node_25, bl_node_44, bl_node_27, bl_node_28, bl_node_34, bl_node_30, bl_node_31, bl_node_32, bl_node_36, bl_node_34, bl_node_35, bl_node_36, bl_node_37, bl_node_38, bl_node_39, bl_node_40, bl_node_41, bl_node_42, bl_node_43, bl_node_44, bl_node_45, bl_node_46, bl_node_47, bl_node_48, bl_node_49, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 2  g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_0, bl_node_1, bl_node_2, bl_node_3, bl_node_4, bl_node_5, bl_node_6, bl_node_7, bl_node_8, bl_node_9, bl_node_10, bl_node_11, bl_node_12, bl_node_13, bl_node_14, bl_node_15, bl_node_16, bl_node_17, bl_node_13, bl_node_19, bl_node_20, bl_node_7, bl_node_22, bl_node_23, bl_node_24, bl_node_25, bl_node_44, bl_node_27, bl_node_28, bl_node_34, bl_node_44, bl_node_34, bl_node_32, bl_node_36, bl_node_34, bl_node_35, bl_node_36, bl_node_37, bl_node_38, bl_node_39, bl_node_40, bl_node_41, bl_node_42, bl_node_43, bl_node_44, bl_node_45, bl_node_47, bl_node_47, bl_node_41, bl_node_27, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 3   g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_0, bl_node_1, bl_node_2, bl_node_3, bl_node_4, bl_node_5, bl_node_6, bl_node_7, bl_node_8, bl_node_9, bl_node_10, bl_node_11, bl_node_12, bl_node_13, bl_node_14, bl_node_15, bl_node_16, bl_node_0, bl_node_13, bl_node_19, bl_node_20, bl_node_7, bl_node_22, bl_node_23, bl_node_24, bl_node_25, bl_node_44, bl_node_27, bl_node_28, bl_node_47, bl_node_44, bl_node_47, bl_node_32, bl_node_36, bl_node_47, bl_node_35, bl_node_36, bl_node_41, bl_node_38, bl_node_35, bl_node_40, bl_node_41, bl_node_42, bl_node_43, bl_node_44, bl_node_45, bl_node_47, bl_node_47, bl_node_41, bl_node_27, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 4    g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_0, bl_node_1, bl_node_2, bl_node_3, bl_node_4, bl_node_5, bl_node_6, bl_node_7, bl_node_8, bl_node_9, bl_node_10, bl_node_11, bl_node_12, bl_node_13, bl_node_8, bl_node_15, bl_node_16, bl_node_0, bl_node_13, bl_node_19, bl_node_20, bl_node_7, bl_node_22, bl_node_10, bl_node_24, bl_node_25, bl_node_44, bl_node_47, bl_node_43, bl_node_47, bl_node_44, bl_node_47, bl_node_32, bl_node_36, bl_node_47, bl_node_35, bl_node_36, bl_node_41, bl_node_38, bl_node_35, bl_node_40, bl_node_41, bl_node_42, bl_node_43, bl_node_44, bl_node_45, bl_node_47, bl_node_47, bl_node_41, bl_node_47, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 5    g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_0, bl_node_1, bl_node_2, bl_node_3, bl_node_4, bl_node_5, bl_node_6, bl_node_7, bl_node_8, bl_node_9, bl_node_10, bl_node_11, bl_node_12, bl_node_13, bl_node_8, bl_node_15, bl_node_16, bl_node_0, bl_node_13, bl_node_19, bl_node_20, bl_node_7, bl_node_22, bl_node_10, bl_node_24, bl_node_25, bl_node_44, bl_node_47, bl_node_43, bl_node_47, bl_node_44, bl_node_47, bl_node_32, bl_node_36, bl_node_47, bl_node_47, bl_node_36, bl_node_41, bl_node_25, bl_node_47, bl_node_40, bl_node_41, bl_node_44, bl_node_43, bl_node_44, bl_node_45, bl_node_47, bl_node_47, bl_node_41, bl_node_47, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 6    g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_0, bl_node_1, bl_node_2, bl_node_3, bl_node_4, bl_node_5, bl_node_6, bl_node_7, bl_node_8, bl_node_9, bl_node_10, bl_node_11, bl_node_12, bl_node_13, bl_node_8, bl_node_15, bl_node_16, bl_node_0, bl_node_13, bl_node_19, bl_node_20, bl_node_7, bl_node_22, bl_node_10, bl_node_24, bl_node_25, bl_node_44, bl_node_47, bl_node_40, bl_node_47, bl_node_44, bl_node_47, bl_node_45, bl_node_47, bl_node_47, bl_node_47, bl_node_47, bl_node_41, bl_node_25, bl_node_47, bl_node_40, bl_node_41, bl_node_44, bl_node_40, bl_node_44, bl_node_45, bl_node_47, bl_node_47, bl_node_41, bl_node_47, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 7     g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_0, bl_node_1, bl_node_2, bl_node_1, bl_node_4, bl_node_5, bl_node_6, bl_node_7, bl_node_8, bl_node_9, bl_node_10, bl_node_11, bl_node_12, bl_node_13, bl_node_8, bl_node_16, bl_node_16, bl_node_0, bl_node_13, bl_node_19, bl_node_20, bl_node_7, bl_node_4, bl_node_10, bl_node_24, bl_node_25, bl_node_44, bl_node_47, bl_node_40, bl_node_47, bl_node_44, bl_node_47, bl_node_45, bl_node_47, bl_node_47, bl_node_47, bl_node_47, bl_node_41, bl_node_25, bl_node_47, bl_node_40, bl_node_41, bl_node_44, bl_node_40, bl_node_44, bl_node_45, bl_node_47, bl_node_47, bl_node_41, bl_node_47, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 8      g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_0, bl_node_1, bl_node_2, bl_node_1, bl_node_4, bl_node_5, bl_node_6, bl_node_7, bl_node_8, bl_node_9, bl_node_10, bl_node_11, bl_node_12, bl_node_13, bl_node_8, bl_node_16, bl_node_16, bl_node_0, bl_node_13, bl_node_13, bl_node_20, bl_node_7, bl_node_4, bl_node_10, bl_node_1, bl_node_25, bl_node_44, bl_node_47, bl_node_40, bl_node_47, bl_node_44, bl_node_47, bl_node_41, bl_node_47, bl_node_47, bl_node_47, bl_node_47, bl_node_41, bl_node_25, bl_node_47, bl_node_40, bl_node_41, bl_node_44, bl_node_40, bl_node_44, bl_node_41, bl_node_47, bl_node_47, bl_node_41, bl_node_47, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 9         g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_0, bl_node_1, bl_node_2, bl_node_1, bl_node_4, bl_node_6, bl_node_6, bl_node_7, bl_node_8, bl_node_9, bl_node_10, bl_node_11, bl_node_12, bl_node_9, bl_node_8, bl_node_16, bl_node_16, bl_node_0, bl_node_9, bl_node_9, bl_node_20, bl_node_7, bl_node_4, bl_node_10, bl_node_1, bl_node_25, bl_node_44, bl_node_47, bl_node_40, bl_node_47, bl_node_44, bl_node_47, bl_node_41, bl_node_47, bl_node_47, bl_node_47, bl_node_47, bl_node_41, bl_node_25, bl_node_47, bl_node_40, bl_node_41, bl_node_44, bl_node_40, bl_node_44, bl_node_41, bl_node_47, bl_node_47, bl_node_41, bl_node_47, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 10          g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_0, bl_node_1, bl_node_2, bl_node_1, bl_node_4, bl_node_6, bl_node_6, bl_node_7, bl_node_8, bl_node_9, bl_node_10, bl_node_11, bl_node_8, bl_node_9, bl_node_8, bl_node_16, bl_node_16, bl_node_0, bl_node_9, bl_node_9, bl_node_4, bl_node_7, bl_node_4, bl_node_10, bl_node_1, bl_node_25, bl_node_44, bl_node_47, bl_node_40, bl_node_47, bl_node_44, bl_node_47, bl_node_41, bl_node_47, bl_node_47, bl_node_47, bl_node_47, bl_node_41, bl_node_25, bl_node_47, bl_node_40, bl_node_41, bl_node_44, bl_node_40, bl_node_44, bl_node_41, bl_node_47, bl_node_47, bl_node_41, bl_node_47, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 11          g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_0, bl_node_1, bl_node_6, bl_node_1, bl_node_4, bl_node_6, bl_node_6, bl_node_7, bl_node_8, bl_node_9, bl_node_10, bl_node_11, bl_node_8, bl_node_9, bl_node_8, bl_node_16, bl_node_16, bl_node_0, bl_node_9, bl_node_9, bl_node_4, bl_node_7, bl_node_4, bl_node_10, bl_node_1, bl_node_40, bl_node_44, bl_node_47, bl_node_40, bl_node_47, bl_node_44, bl_node_47, bl_node_41, bl_node_47, bl_node_47, bl_node_47, bl_node_47, bl_node_41, bl_node_40, bl_node_47, bl_node_40, bl_node_41, bl_node_44, bl_node_40, bl_node_44, bl_node_41, bl_node_47, bl_node_47, bl_node_41, bl_node_47, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 12       g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_0, bl_node_16, bl_node_6, bl_node_16, bl_node_4, bl_node_6, bl_node_6, bl_node_7, bl_node_8, bl_node_9, bl_node_10, bl_node_11, bl_node_8, bl_node_9, bl_node_8, bl_node_16, bl_node_16, bl_node_0, bl_node_9, bl_node_9, bl_node_4, bl_node_7, bl_node_4, bl_node_10, bl_node_16, bl_node_40, bl_node_40, bl_node_47, bl_node_40, bl_node_47, bl_node_40, bl_node_47, bl_node_41, bl_node_47, bl_node_47, bl_node_47, bl_node_47, bl_node_41, bl_node_40, bl_node_47, bl_node_40, bl_node_41, bl_node_40, bl_node_40, bl_node_40, bl_node_41, bl_node_47, bl_node_47, bl_node_41, bl_node_47, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 13          g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_0, bl_node_16, bl_node_6, bl_node_16, bl_node_4, bl_node_6, bl_node_6, bl_node_7, bl_node_8, bl_node_9, bl_node_7, bl_node_6, bl_node_8, bl_node_9, bl_node_8, bl_node_16, bl_node_16, bl_node_0, bl_node_9, bl_node_9, bl_node_4, bl_node_7, bl_node_4, bl_node_7, bl_node_16, bl_node_40, bl_node_40, bl_node_47, bl_node_40, bl_node_47, bl_node_40, bl_node_47, bl_node_41, bl_node_47, bl_node_47, bl_node_47, bl_node_47, bl_node_41, bl_node_40, bl_node_47, bl_node_40, bl_node_41, bl_node_40, bl_node_40, bl_node_40, bl_node_41, bl_node_47, bl_node_47, bl_node_41, bl_node_47, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 14       g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_0, bl_node_16, bl_node_6, bl_node_16, bl_node_16, bl_node_6, bl_node_6, bl_node_7, bl_node_8, bl_node_9, bl_node_7, bl_node_6, bl_node_8, bl_node_9, bl_node_8, bl_node_16, bl_node_16, bl_node_0, bl_node_9, bl_node_9, bl_node_16, bl_node_7, bl_node_16, bl_node_7, bl_node_16, bl_node_40, bl_node_40, bl_node_47, bl_node_40, bl_node_47, bl_node_40, bl_node_47, bl_node_41, bl_node_47, bl_node_47, bl_node_47, bl_node_47, bl_node_41, bl_node_40, bl_node_47, bl_node_40, bl_node_41, bl_node_40, bl_node_40, bl_node_40, bl_node_41, bl_node_47, bl_node_47, bl_node_41, bl_node_47, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 15       g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_0, bl_node_16, bl_node_6, bl_node_16, bl_node_16, bl_node_6, bl_node_6, bl_node_7, bl_node_8, bl_node_9, bl_node_7, bl_node_6, bl_node_8, bl_node_9, bl_node_8, bl_node_16, bl_node_16, bl_node_0, bl_node_9, bl_node_9, bl_node_16, bl_node_7, bl_node_16, bl_node_7, bl_node_16, bl_node_41, bl_node_41, bl_node_47, bl_node_41, bl_node_47, bl_node_41, bl_node_47, bl_node_41, bl_node_47, bl_node_47, bl_node_47, bl_node_47, bl_node_41, bl_node_41, bl_node_47, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_47, bl_node_47, bl_node_41, bl_node_47, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 16    g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_0, bl_node_16, bl_node_6, bl_node_16, bl_node_16, bl_node_6, bl_node_6, bl_node_7, bl_node_16, bl_node_9, bl_node_7, bl_node_6, bl_node_16, bl_node_9, bl_node_16, bl_node_16, bl_node_16, bl_node_0, bl_node_9, bl_node_9, bl_node_16, bl_node_7, bl_node_16, bl_node_7, bl_node_16, bl_node_41, bl_node_41, bl_node_47, bl_node_41, bl_node_47, bl_node_41, bl_node_47, bl_node_41, bl_node_47, bl_node_47, bl_node_47, bl_node_47, bl_node_41, bl_node_41, bl_node_47, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_47, bl_node_47, bl_node_41, bl_node_47, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 17    g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_7, bl_node_16, bl_node_6, bl_node_16, bl_node_16, bl_node_6, bl_node_6, bl_node_7, bl_node_16, bl_node_9, bl_node_7, bl_node_6, bl_node_16, bl_node_9, bl_node_16, bl_node_16, bl_node_16, bl_node_7, bl_node_9, bl_node_9, bl_node_16, bl_node_7, bl_node_16, bl_node_7, bl_node_16, bl_node_41, bl_node_41, bl_node_47, bl_node_41, bl_node_47, bl_node_41, bl_node_47, bl_node_41, bl_node_47, bl_node_47, bl_node_47, bl_node_47, bl_node_41, bl_node_41, bl_node_47, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_47, bl_node_47, bl_node_41, bl_node_47, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 18               g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_7, bl_node_6, bl_node_6, bl_node_6, bl_node_6, bl_node_6, bl_node_6, bl_node_7, bl_node_6, bl_node_9, bl_node_7, bl_node_6, bl_node_6, bl_node_9, bl_node_6, bl_node_6, bl_node_6, bl_node_7, bl_node_9, bl_node_9, bl_node_6, bl_node_7, bl_node_6, bl_node_7, bl_node_6, bl_node_41, bl_node_41, bl_node_47, bl_node_41, bl_node_47, bl_node_41, bl_node_47, bl_node_41, bl_node_47, bl_node_47, bl_node_47, bl_node_47, bl_node_41, bl_node_41, bl_node_47, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_47, bl_node_47, bl_node_41, bl_node_47, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 19               g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_7, bl_node_6, bl_node_6, bl_node_6, bl_node_6, bl_node_6, bl_node_6, bl_node_7, bl_node_6, bl_node_9, bl_node_7, bl_node_6, bl_node_6, bl_node_9, bl_node_6, bl_node_6, bl_node_6, bl_node_7, bl_node_9, bl_node_9, bl_node_6, bl_node_7, bl_node_6, bl_node_7, bl_node_6, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 20               g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_7, bl_node_6, bl_node_6, bl_node_6, bl_node_6, bl_node_6, bl_node_6, bl_node_7, bl_node_6, bl_node_7, bl_node_7, bl_node_6, bl_node_6, bl_node_7, bl_node_6, bl_node_6, bl_node_6, bl_node_7, bl_node_7, bl_node_7, bl_node_6, bl_node_7, bl_node_6, bl_node_7, bl_node_6, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 21               g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, bl_node_41, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 22                                        g1_1, g1_2, g1_3, g1_4, g1_5, g1_6, g1_7, g1_8, g1_9, g1_10, g1_11, g1_12, g1_13, g1_14, g1_15, g1_16, g1_17, g1_18, g1_19, g1_20, g1_21, g1_22, g1_23, g1_24, g1_25, g2_1, g2_2, g2_3, g2_4, g2_5, g2_6, g2_7, g2_8, g2_9, g2_10, g2_11, g2_12, g2_13, g2_14, g2_15, g2_16, g2_17, g2_18, g2_19, g2_20, g2_21, g2_22, g2_23, g2_24, g2_25, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, node, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, bl_node_7, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0