Dump state of current SBM model to dataframe

Exports the current state of the network as dataframe with each node as a row and columns for node id, parent id, node type, and node level. Can also be accomplished by attr(sbm, 'state').

get_state(x)

Value

A dataframe describing the current state of the network:

id

id of node

parent

id of node's parent. none if node has no parent.

type

type of node

level

level of node in network hierarchy

See also

Examples

# Small edge list to build network
edges <- dplyr::tribble(
  ~from, ~to,
  "a1"   , "b1"   ,
  "a1"   , "b2"   ,
  "a1"   , "b3"   ,
  "a2"   , "b1"   ,
  "a2"   , "b4"   ,
  "a3"   , "b1"
)

# A small simulated network with no blocks
net <- new_sbm_network(edges)

get_state(net)

#>   id parent type level
#> 1 a1   none node     0
#> 2 a2   none node     0
#> 3 a3   none node     0
#> 4 b1   none node     0
#> 5 b2   none node     0
#> 6 b3   none node     0
#> 7 b4   none node     0

# Add some blocks and state will reflect
net %>% initialize_blocks(2) %>% get_state()

#>      id parent type level
#> 1    a1  0-1_0 node     0
#> 2    a2  0-1_0 node     0
#> 3    a3  0-1_0 node     0
#> 4    b1  0-1_1 node     0
#> 5    b2  0-1_0 node     0
#> 6    b3  0-1_1 node     0
#> 7    b4  0-1_1 node     0
#> 8 0-1_0   none node     1
#> 9 0-1_1   none node     1