Assign groups for all nodes

Initializes groups for a given level of network. When num_groups = -1 (the default) this will simply give every node its own group. Otherwise it will randomly draw each nodes group from num_group possible parent nodes for each node type. E.g. if network is bipartite and num_groups == 3 a total of 3*2 = 6 groups will be added and assigned randomly to nodes. Because of random assignment of parents there may sometimes be less groups than num_groups * <num_types> due to some group nodes never being assigned to a child node.

initialize_groups(sbm, num_groups = -1, level = 0)

Arguments

sbm
SBM model object as created by create_sbm().

num_groups
Total number of groups to initialize for each unique node type. If set to -1, then a single group is made for every node.

level
Level of child node. This will almost always stay at default value of 0.

Value

SBM model object modified with new group nodes and node hierarchy modified. Note that object is modified in place as well.

Examples

# Helper function to get number of groups in a model
get_num_groups <- function(sbm){sum(get_state(sbm)$level == 1)}

# Initialize a simple bipartite network with 6 total nodes (3 of each type)
my_nodes <- dplyr::tibble(
  id = c("a1", "a2", "a3", "b1", "b2", "b3"),
  type = c("a", "a", "a", "b", "b", "b")
)
my_sbm <- create_sbm(nodes = my_nodes)

# Default values of function will give every node its own group
my_sbm %>% initialize_groups()

#>       id parent type level
#> 1     a1  0-1_0    a     0
#> 2     a2  0-1_1    a     0
#> 3     a3  0-1_2    a     0
#> 4     b1  1-1_3    b     0
#> 5     b2  1-1_4    b     0
#> 6     b3  1-1_5    b     0
#> 7  0-1_0   none    a     1
#> 8  0-1_1   none    a     1
#> 9  0-1_2   none    a     1
#> 10 1-1_3   none    b     1
#> 11 1-1_4   none    b     1
#> 12 1-1_5   none    b     1

get_num_groups(my_sbm)

#> [1] 6

# You can also decide to have a given number of groups randomly assigned Here
# four groups result because two random groups are made for each of the two
# types
my_sbm %>% initialize_groups(num_groups = 2)

#>       id parent type level
#> 1     a1  0-1_0    a     0
#> 2     a2  0-1_0    a     0
#> 3     a3  0-1_0    a     0
#> 4     b1  1-1_3    b     0
#> 5     b2  1-1_2    b     0
#> 6     b3  1-1_3    b     0
#> 7  0-1_0   none    a     1
#> 8  0-1_1   none    a     1
#> 9  1-1_2   none    b     1
#> 10 1-1_3   none    b     1

get_num_groups(my_sbm)

#> [1] 4