Run a single MCMC sweep over nodes

Runs a single Metropolis Hastings MCMC sweep across all nodes at specified level (For algorithm details see Piexoto, 2018). Each node is given a chance to move blocks or stay in current block and all nodes are processed in random order. Takes the level that the sweep should take place on (int) and if new blocks blocks can be proposed and empty blocks removed (boolean).

mcmc_sweep(
  sbm,
  num_sweeps = 1,
  eps = 0.1,
  variable_num_blocks = TRUE,
  track_pairs = FALSE,
  level = 0,
  verbose = FALSE,
  return_sbm_network = TRUE
)

Arguments

sbm
sbm_network object as created by new_sbm_network().

num_sweeps
Number of times all nodes are passed through for move proposals.

eps
Controls randomness of move proposals. Effects both the block merging and mcmc sweeps.

variable_num_blocks
Should the model allow new blocks to be created or empty blocks removed while sweeping or should number of blocks remain constant?

track_pairs
Return a dataframe with all pairs of nodes along with the number of sweeps they shared the same group?

level
Level of nodes who's blocks will have their block membership run through MCMC proposal-accept routine.

verbose
If set to TRUE then each proposed move for all sweeps will have information given on entropy delta, probability of moving, and if the move were accepted printed to the console.

Value

An S3 object of class sbm_network. For details see new_sbm_network section "Class structure."

See also

Examples

set.seed(42)

# Start with a small simulated network with random block assignments
net <- sim_basic_block_network(n_blocks = 4, n_nodes_per_block = 15) %>%
  initialize_blocks(num_blocks = 4)

# Calculate entropy with random blocks
get_entropy(net)

#> [1] 1283.42

# Run some MCMC sweeps
net <- mcmc_sweep(net, num_sweeps = 25, variable_num_blocks = FALSE)

# Entropy after sweeps
get_entropy(net)

#> [1] 1282.171

# Per-sweep level information
get_sweep_results(net)

#> $nodes_moved
#>   [1] "g4_12" "g4_11" "g1_5"  "g3_4"  "g4_14" "g3_8"  "g1_9"  "g1_3"  "g4_3" 
#>  [10] "g4_15" "g3_6"  "g1_2"  "g1_13" "g1_6"  "g3_9"  "g2_11" "g2_1"  "g2_13"
#>  [19] "g3_11" "g3_13" "g2_14" "g3_15" "g2_15" "g3_5"  "g2_10" "g3_3"  "g2_2" 
#>  [28] "g1_14" "g3_10" "g4_1"  "g2_3"  "g2_12" "g2_6"  "g4_6"  "g1_10" "g1_1" 
#>  [37] "g4_7"  "g1_12" "g3_5"  "g4_1"  "g2_9"  "g1_15" "g1_1"  "g1_11" "g1_2" 
#>  [46] "g4_13" "g2_10" "g3_15" "g1_9"  "g1_7"  "g3_4"  "g2_5"  "g4_3"  "g1_8" 
#>  [55] "g4_6"  "g1_4"  "g2_2"  "g3_11" "g2_14" "g2_11" "g3_2"  "g3_9"  "g2_12"
#>  [64] "g3_6"  "g1_12" "g1_10" "g2_14" "g3_4"  "g1_3"  "g3_1"  "g3_2"  "g1_12"
#>  [73] "g3_12" "g3_13" "g1_2"  "g2_11" "g3_15" "g2_4"  "g1_13" "g4_13" "g2_7" 
#>  [82] "g1_7"  "g2_9"  "g2_15" "g2_3"  "g1_10" "g3_6"  "g2_8"  "g3_10" "g2_1" 
#>  [91] "g1_4"  "g2_10" "g4_8"  "g2_5"  "g4_4"  "g4_14" "g3_5"  "g1_6"  "g1_13"
#> [100] "g4_15" "g2_5"  "g2_9"  "g3_8"  "g2_10" "g3_9"  "g4_8"  "g1_5"  "g2_6" 
#> [109] "g4_2"  "g3_6"  "g3_11" "g3_7"  "g4_5"  "g1_15" "g2_4"  "g3_4"  "g1_6" 
#> [118] "g1_4"  "g1_3"  "g2_11" "g1_10" "g2_15" "g1_12" "g2_12" "g1_8"  "g1_5" 
#> [127] "g4_13" "g4_12" "g3_3"  "g1_9"  "g2_7"  "g3_1"  "g3_14" "g2_14" "g3_5" 
#> [136] "g2_10" "g3_13" "g2_13" "g3_9"  "g4_3"  "g1_2"  "g2_4"  "g3_15" "g4_8" 
#> [145] "g4_4"  "g3_10" "g3_4"  "g4_11" "g2_5"  "g2_3"  "g4_9"  "g1_3"  "g2_14"
#> [154] "g4_10" "g2_11" "g4_4"  "g4_8"  "g2_6"  "g4_5"  "g4_7"  "g3_12" "g2_12"
#> [163] "g4_13" "g1_1"  "g1_12" "g4_6"  "g1_6"  "g4_12" "g1_2"  "g3_13" "g4_1" 
#> [172] "g4_3"  "g3_11" "g1_8"  "g3_9"  "g2_5"  "g3_10" "g4_2"  "g3_5"  "g4_11"
#> [181] "g1_14" "g1_15" "g1_11" "g2_8"  "g3_15" "g3_6"  "g1_2"  "g1_14" "g2_5" 
#> [190] "g1_7"  "g1_8"  "g3_9"  "g2_12" "g2_11" "g4_1"  "g1_10" "g1_1"  "g2_6" 
#> [199] "g1_5"  "g4_5"  "g4_4"  "g4_15" "g2_3"  "g3_14" "g4_9"  "g3_1"  "g4_6" 
#> [208] "g3_5"  "g3_3"  "g4_2"  "g2_14" "g1_9"  "g2_2"  "g4_12" "g3_8"  "g3_13"
#> [217] "g4_13" "g1_3"  "g3_2"  "g4_3"  "g4_11" "g2_8"  "g1_13" "g1_12" "g3_10"
#> [226] "g1_14" "g2_5"  "g1_11" "g4_1"  "g4_2"  "g2_9"  "g3_5"  "g2_11" "g1_9" 
#> [235] "g2_13" "g4_6"  "g4_12" "g3_14" "g2_7"  "g4_5"  "g1_2"  "g3_4"  "g2_10"
#> [244] "g4_15" "g4_4"  "g3_1"  "g1_7"  "g4_11" "g2_12" "g4_3"  "g4_14" "g2_3" 
#> [253] "g1_10" "g3_9"  "g1_1"  "g4_10" "g3_11" "g2_4"  "g2_1"  "g1_13" "g4_9" 
#> [262] "g3_5"  "g3_12" "g4_10" "g2_10" "g2_5"  "g3_9"  "g4_5"  "g4_7"  "g2_4" 
#> [271] "g1_10" "g4_1"  "g1_2"  "g2_9"  "g3_15" "g4_9"  "g2_15" "g2_11" "g1_6" 
#> [280] "g4_4"  "g3_6"  "g3_1"  "g1_1"  "g1_11" "g2_3"  "g4_14" "g4_2"  "g3_13"
#> [289] "g3_2"  "g2_1"  "g1_5"  "g4_12" "g2_7"  "g4_6"  "g3_11" "g1_6"  "g1_14"
#> [298] "g1_9"  "g1_2"  "g4_9"  "g1_12" "g2_10" "g3_7"  "g3_1"  "g3_6"  "g3_10"
#> [307] "g1_1"  "g4_3"  "g1_11" "g4_14" "g4_10" "g3_5"  "g2_1"  "g4_2"  "g2_2" 
#> [316] "g2_15" "g2_12" "g4_7"  "g3_8"  "g2_14" "g3_13" "g3_12" "g4_1"  "g3_3" 
#> [325] "g2_5"  "g2_11" "g2_6"  "g3_15" "g1_3"  "g4_4"  "g3_4"  "g2_13" "g2_4" 
#> [334] "g4_5"  "g1_10" "g4_2"  "g2_15" "g2_13" "g2_3"  "g1_13" "g4_8"  "g1_14"
#> [343] "g1_1"  "g4_3"  "g3_6"  "g4_9"  "g2_14" "g3_12" "g1_11" "g2_10" "g3_3" 
#> [352] "g1_7"  "g2_6"  "g1_3"  "g3_15" "g1_8"  "g1_12" "g4_13" "g2_12" "g4_14"
#> [361] "g3_14" "g1_15" "g1_2"  "g3_8"  "g2_2"  "g3_14" "g4_14" "g4_2"  "g3_1" 
#> [370] "g4_6"  "g2_4"  "g2_12" "g4_12" "g2_10" "g4_7"  "g1_13" "g1_11" "g3_3" 
#> [379] "g1_4"  "g2_5"  "g3_8"  "g4_8"  "g1_12" "g2_9"  "g2_15" "g1_8"  "g1_10"
#> [388] "g2_3"  "g1_9"  "g4_1"  "g3_11" "g1_3"  "g2_7"  "g3_2"  "g3_6"  "g3_7" 
#> [397] "g1_15" "g4_11" "g4_3"  "g4_4"  "g4_5"  "g2_8"  "g3_6"  "g1_7"  "g4_9" 
#> [406] "g1_4"  "g4_11" "g1_8"  "g3_11" "g2_2"  "g4_7"  "g3_9"  "g3_2"  "g3_14"
#> [415] "g2_9"  "g4_2"  "g4_8"  "g3_3"  "g3_7"  "g2_10" "g4_10" "g3_5"  "g1_1" 
#> [424] "g2_1"  "g3_8"  "g2_3"  "g2_8"  "g4_3"  "g3_1"  "g4_1"  "g3_13" "g1_9" 
#> [433] "g4_14" "g1_12" "g1_2"  "g1_13" "g4_7"  "g2_5"  "g3_5"  "g4_9"  "g2_2" 
#> [442] "g3_8"  "g3_4"  "g4_4"  "g1_12" "g3_2"  "g2_12" "g3_11" "g4_2"  "g4_10"
#> [451] "g3_1"  "g1_1"  "g4_11" "g2_3"  "g4_1"  "g3_14" "g2_9"  "g1_15" "g2_8" 
#> [460] "g4_5"  "g1_9"  "g2_14" "g1_7"  "g2_10" "g3_3"  "g1_14" "g4_8"  "g1_6" 
#> [469] "g1_7"  "g1_2"  "g1_12" "g2_13" "g4_2"  "g2_3"  "g1_13" "g1_1"  "g1_10"
#> [478] "g2_15" "g2_10" "g4_6"  "g3_11" "g2_2"  "g3_10" "g4_9"  "g2_6"  "g3_7" 
#> [487] "g4_10" "g3_3"  "g1_8"  "g4_3"  "g3_13" "g2_7"  "g3_1"  "g3_15" "g4_12"
#> [496] "g1_15" "g4_8"  "g3_8"  "g2_4"  "g2_14" "g3_9"  "g1_12" "g3_6"  "g1_2" 
#> [505] "g4_14" "g1_14" "g1_9"  "g2_13" "g3_15" "g4_8"  "g3_4"  "g2_5"  "g2_6" 
#> [514] "g4_4"  "g4_15" "g3_10" "g1_7"  "g2_2"  "g2_10" "g4_11" "g1_8"  "g3_11"
#> [523] "g3_12" "g3_2"  "g2_1"  "g4_9"  "g3_8"  "g3_9"  "g1_10" "g4_5"  "g2_3" 
#> [532] "g4_7"  "g2_9"  "g2_14" "g1_3"  "g1_5"  "g2_8"  "g4_10" "g1_13" "g4_13"
#> [541] "g3_3"  "g1_4"  "g3_5"  "g1_11" "g2_5"  "g1_10" "g3_7"  "g1_2"  "g3_11"
#> [550] "g3_5"  "g4_8"  "g2_13" "g1_6"  "g1_15" "g4_5"  "g4_6"  "g3_1"  "g2_8" 
#> [559] "g4_9"  "g1_5"  "g1_12" "g1_9"  "g2_3"  "g1_8"  "g1_1"  "g3_9"  "g1_14"
#> [568] "g4_14" "g2_2"  "g4_3"  "g4_7"  "g4_13" "g4_1"  "g3_8"  "g3_15" "g3_6" 
#> [577] "g1_13" "g4_3"  "g3_8"  "g3_5"  "g2_5"  "g2_13" "g4_7"  "g2_6"  "g2_12"
#> [586] "g1_2"  "g1_5"  "g4_14" "g4_2"  "g3_4"  "g2_3"  "g1_8"  "g2_10" "g1_3" 
#> [595] "g3_1"  "g3_10" "g4_10" "g4_13" "g4_6"  "g2_15" "g2_7"  "g4_12" "g3_6" 
#> [604] "g3_14" "g1_12" "g2_4"  "g2_2"  "g1_9"  "g3_12" "g3_7"  "g1_7"  "g3_15"
#> [613] "g2_1"  "g1_1"  "g1_6"  "g4_4"  "g2_12" "g4_2"  "g2_5"  "g1_15" "g1_6" 
#> [622] "g3_5"  "g4_13" "g4_14" "g1_4"  "g1_10" "g3_10" "g3_11" "g1_11" "g1_12"
#> [631] "g2_4"  "g2_3"  "g1_8"  "g1_13" "g2_14" "g4_5"  "g2_2"  "g2_9"  "g2_13"
#> [640] "g3_8"  "g3_2"  "g4_6"  "g3_6"  "g1_1"  "g3_9"  "g2_15" "g2_11" "g4_4" 
#> [649] "g3_14" "g1_9"  "g3_7"  "g4_8"  "g3_15" "g1_14" "g2_9"  "g4_11" "g1_13"
#> [658] "g4_2"  "g2_6"  "g3_12" "g1_15" "g1_8"  "g1_11" "g3_10" "g1_1"  "g4_14"
#> [667] "g1_2"  "g3_11" "g3_6"  "g4_4"  "g3_4"  "g2_3"  "g1_4"  "g4_8"  "g3_14"
#> [676] "g2_11" "g4_10" "g1_9"  "g3_5"  "g3_9"  "g3_1"  "g1_5"  "g2_5"  "g2_15"
#> [685] "g1_7"  "g4_7"  "g3_15" "g3_2"  "g1_1"  "g4_10" "g3_14" "g1_12" "g2_10"
#> [694] "g4_9"  "g2_1"  "g3_1"  "g2_3"  "g3_9"  "g3_11" "g4_13" "g4_6"  "g1_4" 
#> [703] "g1_3"  "g3_8"  "g4_14" "g2_11" "g4_12" "g3_6"  "g1_9"  "g1_5"  "g3_13"
#> [712] "g3_5"  "g1_15" "g3_7"  "g2_13" "g2_9"  "g4_15" "g3_12" "g2_4"  "g3_9" 
#> [721] "g4_4"  "g1_1"  "g3_1"  "g3_10" "g2_10" "g4_13" "g1_3"  "g1_9"  "g2_3" 
#> [730] "g2_6"  "g4_1"  "g3_12" "g3_3"  "g1_10" "g3_15" "g3_7"  "g2_8"  "g2_13"
#> [739] "g2_15" "g2_14" "g2_2"  "g1_13" "g1_4"  "g2_12" "g3_2"  "g1_12" "g3_6" 
#> [748] "g1_7"  "g1_15" "g3_4"  "g4_9"  "g3_14" "g4_14" "g3_12" "g2_6"  "g4_12"
#> [757] "g2_14" "g4_8"  "g4_13" "g2_4"  "g1_2"  "g4_2"  "g1_8"  "g1_7"  "g3_8" 
#> [766] "g3_5"  "g1_15" "g4_10" "g4_11" "g4_3"  "g4_6"  "g1_4"  "g1_3"  "g2_12"
#> [775] "g2_1"  "g4_9"  "g2_5"  "g3_2"  "g3_7"  "g2_15" "g3_4"  "g3_11" "g2_13"
#> [784] "g2_9"  "g2_1"  "g3_8"  "g3_4"  "g3_6"  "g4_11" "g1_14" "g3_14" "g3_11"
#> [793] "g2_8"  "g4_10" "g2_7"  "g4_9"  "g2_10" "g4_12" "g4_8"  "g4_5"  "g1_6" 
#> [802] "g1_4"  "g4_3"  "g4_2"  "g2_12" "g1_10" "g2_5"  "g3_13" "g4_14" "g1_1" 
#> [811] "g4_15" "g1_7"  "g4_1"  "g3_3"  "g3_12" "g4_15" "g4_12" "g3_8"  "g1_5" 
#> [820] "g4_6"  "g2_10" "g1_8"  "g2_4"  "g1_9"  "g3_7"  "g2_12" "g3_10" "g4_1" 
#> [829] "g4_5"  "g3_11" "g1_7"  "g3_6"  "g4_9"  "g2_2"  "g2_6"  "g4_8"  "g4_3" 
#> [838] "g2_11" "g4_10" "g4_11" "g1_1"  "g1_6"  "g1_10" "g1_13" "g4_7"  "g2_15"
#> 
#> $sweep_info
#>    entropy_delta num_nodes_moved
#> 1    -2.33393121              38
#> 2    -0.61595700              28
#> 3    -2.63906596              32
#> 4     2.13671199              26
#> 5     1.18443210              25
#> 6     2.81419930              37
#> 7    -2.18930449              39
#> 8     0.43359826              36
#> 9     0.37297737              34
#> 10    0.33336536              37
#> 11   -2.62058245              33
#> 12    2.59976824              37
#> 13   -2.06319466              32
#> 14   -0.14369534              33
#> 15    2.96329957              34
#> 16    0.01143795              42
#> 17   -2.61414595              33
#> 18   -0.17141884              40
#> 19   -0.99874711              37
#> 20   -0.49692188              31
#> 21   -0.48848022              35
#> 22    2.82962536              32
#> 23   -0.71226268              32
#> 24    1.81990144              32
#> 25   -0.66083977              31
#> 

# Use track_pairs = TRUE to get an idea of node-pair similarity by looking at
# how often every pair of nodes is connected over sweeps
net %>%
  mcmc_sweep(num_sweeps = 25, track_pairs = TRUE) %>%
  get_sweep_pair_counts()

#>     node_a node_b times_connected proportion_connected
#> 1     g4_9   g4_8               3                 0.12
#> 2     g4_9   g4_7              19                 0.76
#> 3     g4_8   g4_7              18                 0.72
#> 4     g4_9   g4_6              19                 0.76
#> 5     g4_7   g4_6               2                 0.08
#> 6     g4_7   g4_5              16                 0.64
#> 7     g4_9   g4_4              19                 0.76
#> 8     g4_8   g4_4              12                 0.48
#> 9     g4_6   g4_4               2                 0.08
#> 10    g4_8   g4_3               9                 0.36
#> 11    g4_6   g4_3               7                 0.28
#> 12    g4_4   g4_3               4                 0.16
#> 13    g4_9   g4_2              22                 0.88
#> 14    g4_8   g4_2              15                 0.60
#> 15    g4_7   g4_2              11                 0.44
#> 16    g4_6   g4_2               3                 0.12
#> 17    g4_5   g4_2              15                 0.60
#> 18    g4_9   g4_3               3                 0.12
#> 19    g4_9  g4_15              16                 0.64
#> 20    g4_8  g4_15               1                 0.04
#> 21    g4_7  g4_15              14                 0.56
#> 22    g4_4  g4_15              17                 0.68
#> 23    g4_9  g4_14               4                 0.16
#> 24    g4_8  g4_14              14                 0.56
#> 25    g4_7  g4_14               8                 0.32
#> 26    g4_6  g4_14              17                 0.68
#> 27    g4_5  g4_14              14                 0.56
#> 28    g4_4  g4_14              20                 0.80
#> 29    g4_8  g4_13              11                 0.44
#> 30    g4_7  g4_13              19                 0.76
#> 31    g4_5  g4_13              10                 0.40
#> 32    g4_4  g4_13              11                 0.44
#> 33    g4_3  g4_13              13                 0.52
#> 34    g4_2  g4_13               8                 0.32
#> 35   g4_15  g4_13               5                 0.20
#> 36   g4_14  g4_13              14                 0.56
#> 37    g4_9  g4_12               2                 0.08
#> 38    g4_7  g4_12              12                 0.48
#> 39    g4_4  g4_12              11                 0.44
#> 40    g4_3  g4_12               9                 0.36
#> 41    g4_2  g4_12               9                 0.36
#> 42   g4_15  g4_12               5                 0.20
#> 43   g4_13  g4_12               8                 0.32
#> 44    g4_9  g4_11               4                 0.16
#> 45    g4_7  g4_11              18                 0.72
#> 46    g4_5  g4_11               7                 0.28
#> 47    g4_3  g4_11              10                 0.40
#> 48   g4_15  g4_11              15                 0.60
#> 49   g4_14  g4_11              13                 0.52
#> 50    g4_9  g4_10              12                 0.48
#> 51    g4_5  g4_10              18                 0.72
#> 52    g4_4  g4_10              16                 0.64
#> 53    g4_2  g4_10              14                 0.56
#> 54   g4_13  g4_10              17                 0.68
#> 55   g4_11  g4_10              13                 0.52
#> 56    g4_9   g4_1               9                 0.36
#> 57    g4_8   g4_1               9                 0.36
#> 58    g4_7   g4_1              15                 0.60
#> 59    g4_6   g4_1              10                 0.40
#> 60    g4_5   g4_1              16                 0.64
#> 61    g4_3   g4_1               6                 0.24
#> 62   g4_14   g4_1              17                 0.68
#> 63   g4_13   g4_1              12                 0.48
#> 64   g4_12   g4_1               7                 0.28
#> 65   g4_11   g4_1              17                 0.68
#> 66   g4_10   g4_1              11                 0.44
#> 67    g4_9   g3_9               4                 0.16
#> 68    g4_8   g3_9               7                 0.28
#> 69    g4_5   g3_9              12                 0.48
#> 70    g4_4   g3_9              15                 0.60
#> 71    g4_3   g3_9              17                 0.68
#> 72    g4_2   g3_9              21                 0.84
#> 73   g4_15   g3_9               4                 0.16
#> 74   g4_14   g3_9              13                 0.52
#> 75   g4_13   g3_9              12                 0.48
#> 76   g4_12   g3_9               7                 0.28
#> 77   g4_11   g3_9              16                 0.64
#> 78    g4_1   g3_9               9                 0.36
#> 79    g4_9   g3_8               3                 0.12
#> 80    g4_6   g3_8              12                 0.48
#> 81    g4_4   g3_8              18                 0.72
#> 82    g4_3   g3_8              15                 0.60
#> 83    g4_2   g3_8              13                 0.52
#> 84   g4_15   g3_8               3                 0.12
#> 85   g4_13   g3_8              14                 0.56
#> 86   g4_12   g3_8              15                 0.60
#> 87   g4_11   g3_8              10                 0.40
#> 88    g4_7   g3_7              16                 0.64
#> 89    g4_6   g3_7               6                 0.24
#> 90    g4_7   g3_9               9                 0.36
#> 91    g4_5   g3_7              16                 0.64
#> 92    g4_2   g3_7               6                 0.24
#> 93   g4_13   g3_7              11                 0.44
#> 94   g4_11   g3_7              11                 0.44
#> 95    g4_1   g3_7              11                 0.44
#> 96    g3_8   g3_7              12                 0.48
#> 97   g4_10   g3_9               7                 0.28
#> 98    g4_8   g3_6               5                 0.20
#> 99    g4_7   g3_6              19                 0.76
#> 100   g4_4   g3_6              17                 0.68
#> 101   g4_3   g3_6              17                 0.68
#> 102  g4_15   g3_6              16                 0.64
#> 103  g4_14   g3_6               3                 0.12
#> 104  g4_12   g3_6              11                 0.44
#> 105  g4_11   g3_6              11                 0.44
#> 106   g3_9   g3_6              13                 0.52
#> 107   g3_7   g3_6              11                 0.44
#> 108   g4_9   g3_5               5                 0.20
#> 109   g2_5   g1_5               0                 0.00
#> 110   g4_5   g3_5              17                 0.68
#> 111   g4_4   g3_5              10                 0.40
#> 112   g4_2   g3_5               8                 0.32
#> 113   g3_8   g1_2               0                 0.00
#> 114  g2_10   g1_7              19                 0.76
#> 115  g4_13   g3_5              14                 0.56
#> 116  g4_12   g3_5              13                 0.52
#> 117  g4_11   g3_5              14                 0.56
#> 118   g4_9   g4_5              14                 0.56
#> 119  g4_10   g3_5              13                 0.52
#> 120   g4_1   g3_5              14                 0.56
#> 121   g3_9   g3_5              17                 0.68
#> 122   g3_8   g3_5              13                 0.52
#> 123   g3_6   g3_5               6                 0.24
#> 124  g4_10  g3_13              10                 0.40
#> 125   g4_5   g3_4              19                 0.76
#> 126   g4_4   g3_4               7                 0.28
#> 127  g4_14   g3_4               5                 0.20
#> 128   g1_8  g1_13               8                 0.32
#> 129  g4_12   g3_4              17                 0.68
#> 130   g4_2   g4_1              14                 0.56
#> 131  g4_10   g3_4              15                 0.60
#> 132   g3_9   g3_4              10                 0.40
#> 133   g3_7   g3_4              10                 0.40
#> 134   g4_9   g3_3              21                 0.84
#> 135   g4_8   g3_3               2                 0.08
#> 136   g4_7   g3_3               2                 0.08
#> 137   g3_2  g2_14              16                 0.64
#> 138   g4_5   g3_3               6                 0.24
#> 139   g4_4   g3_3               3                 0.12
#> 140   g4_2   g3_3               2                 0.08
#> 141  g4_15  g1_11              20                 0.80
#> 142   g3_6   g3_3              20                 0.80
#> 143   g4_9   g3_2              22                 0.88
#> 144   g4_7   g3_2               6                 0.24
#> 145   g3_5  g1_10               1                 0.04
#> 146   g3_5   g1_3               0                 0.00
#> 147   g3_2  g2_13               5                 0.20
#> 148   g4_5   g3_2              12                 0.48
#> 149   g4_3   g3_2               9                 0.36
#> 150   g3_4   g3_3              14                 0.56
#> 151   g3_6   g1_4               0                 0.00
#> 152  g4_15   g3_2              16                 0.64
#> 153  g4_14   g3_2               5                 0.20
#> 154  g4_13   g3_2              18                 0.72
#> 155  g4_11   g3_2              15                 0.60
#> 156  g4_10   g3_2              18                 0.72
#> 157   g4_3   g1_6               0                 0.00
#> 158   g3_9   g3_2              14                 0.56
#> 159   g3_3   g3_2              18                 0.72
#> 160   g4_7  g3_15              16                 0.64
#> 161  g4_11   g1_8               1                 0.04
#> 162   g4_4  g3_15              12                 0.48
#> 163   g4_3  g3_15               7                 0.28
#> 164   g4_2  g2_11               5                 0.20
#> 165  g4_11  g2_11               8                 0.32
#> 166   g4_2  g3_15               9                 0.36
#> 167   g3_7  g1_10               0                 0.00
#> 168  g4_15  g3_15               4                 0.16
#> 169   g3_5   g2_3               3                 0.12
#> 170  g4_13  g3_15               5                 0.20
#> 171  g4_11  g3_15               5                 0.20
#> 172   g3_9  g3_15               5                 0.20
#> 173   g3_7  g3_15               5                 0.20
#> 174   g2_3   g1_8              25                 1.00
#> 175   g2_9   g2_8               7                 0.28
#> 176   g3_3  g3_15              21                 0.84
#> 177   g2_1   g1_3               0                 0.00
#> 178  g4_11   g1_3               0                 0.00
#> 179   g4_9  g3_14               1                 0.04
#> 180   g1_2  g1_10              14                 0.56
#> 181   g4_6  g3_14               4                 0.16
#> 182   g4_3  g3_14              19                 0.76
#> 183  g3_13   g2_8              10                 0.40
#> 184   g4_2  g3_14              11                 0.44
#> 185   g3_6  g2_15               8                 0.32
#> 186  g4_12  g3_14               8                 0.32
#> 187  g4_15   g2_1              11                 0.44
#> 188   g3_8   g3_4              18                 0.72
#> 189  g4_11  g3_14               8                 0.32
#> 190  g4_10  g3_14               8                 0.32
#> 191   g4_1  g3_14              17                 0.68
#> 192   g3_9  g3_14              14                 0.56
#> 193   g4_8  g3_13              10                 0.40
#> 194  g2_13  g1_10               0                 0.00
#> 195   g4_7  g3_13              12                 0.48
#> 196   g4_6  g3_13               7                 0.28
#> 197  g3_15   g2_4               4                 0.16
#> 198   g4_5  g3_13              17                 0.68
#> 199   g3_4   g2_5              11                 0.44
#> 200   g2_3  g2_11               8                 0.32
#> 201   g4_5  g3_15              10                 0.40
#> 202   g4_3  g3_13              13                 0.52
#> 203   g4_2  g3_13              12                 0.48
#> 204  g4_15  g3_14              24                 0.96
#> 205   g3_9  g3_13              13                 0.52
#> 206  g4_14  g3_15               9                 0.36
#> 207   g3_3  g1_11               1                 0.04
#> 208   g3_7  g3_13              11                 0.44
#> 209   g2_6  g1_12               0                 0.00
#> 210   g2_6  g1_11               2                 0.08
#> 211   g3_5  g3_13              10                 0.40
#> 212  g4_12  g2_12              13                 0.52
#> 213   g3_2  g3_13               8                 0.32
#> 214   g2_9   g2_1              13                 0.52
#> 215   g4_4   g2_7               8                 0.32
#> 216  g4_13  g3_13              14                 0.56
#> 217  g3_15  g3_13               6                 0.24
#> 218   g4_5  g3_12               5                 0.20
#> 219   g4_4  g3_14               7                 0.28
#> 220   g2_3   g1_6              24                 0.96
#> 221  g4_13  g3_12              12                 0.48
#> 222  g4_14  g2_13              10                 0.40
#> 223   g3_9  g3_12              14                 0.56
#> 224   g3_2  g3_12              11                 0.44
#> 225   g3_3   g1_6               1                 0.04
#> 226  g3_15  g3_12               4                 0.16
#> 227  g3_14  g3_12              15                 0.60
#> 228   g4_8  g3_11              12                 0.48
#> 229   g2_4   g1_1               5                 0.20
#> 230   g4_6  g3_11               1                 0.04
#> 231   g4_2  g4_15               1                 0.04
#> 232   g4_5  g3_11              16                 0.64
#> 233   g2_8   g1_3               2                 0.08
#> 234   g4_3  g3_11              20                 0.80
#> 235  g4_14  g1_12               0                 0.00
#> 236  g4_14  g3_11              20                 0.80
#> 237   g4_1  g3_12              19                 0.76
#> 238  g3_12   g1_6              24                 0.96
#> 239  g3_10   g1_7               2                 0.08
#> 240  g4_13  g3_11              10                 0.40
#> 241   g3_6  g3_13              17                 0.68
#> 242   g4_1   g2_2              15                 0.60
#> 243  g4_11  g3_11              19                 0.76
#> 244   g4_1  g3_11              12                 0.48
#> 245   g3_8  g3_11              14                 0.56
#> 246   g3_7  g3_11               4                 0.16
#> 247   g3_6  g3_11               3                 0.12
#> 248   g1_9  g1_14              13                 0.52
#> 249   g3_5  g3_11              13                 0.52
#> 250  g3_12  g3_11              10                 0.40
#>  [ reached 'max' / getOption("max.print") -- omitted 1520 rows ]