Build new SBM s4 class

Builds an s4 class exported by Rcpp modules. Generally used internally. For making the main sbm_network class, use new_sbm_network instead.

new_sbm_s4(
  nodes,
  edges,
  allowed_edge_types = NULL,
  node_types = NULL,
  random_seed = NULL,
  state = NULL
)

Arguments

nodes
Optional dataframe that links a node id to its type, for when model is polypartite.

edges
Dataframe with a from and two column encoding edges between string node ids (direction does not matter).

allowed_edge_types
A dataframe in the same format of as edges that contains allowed pairs (order does not matter) of possible node type combinations across edges. For instance: nodes of type pollinator are allowed to connect to flower nodes but not to other pollinator nodes. If this is left undefined, it is inferred from edges.

node_types
Character vector with unique node types in network. If not provided, this will be deduced from present types in nodes.

random_seed
Integer seed to be passed to model's internal random sampling engine. Note that if the model is restored from a saved state this seed will be initialized again to the start value which will harm reproducability.

Value

An SBM S4 class

Examples

nodes <- dplyr::tribble(~id,     ~type,
"a1",    "a",
"a2",    "a",
"a3",    "a",
"b1",    "b",
"b2",    "b",
"b3",    "b",
"b4",    "b")

node_types <- c("a", "b")

edges <- dplyr::tribble(~a,     ~b,
                        "a1",    "b1",
                        "a1",    "b2",
                        "a1",    "b3",
                        "a2",    "b1",
                        "a2",    "b4",
                        "a3",    "b1")

random_seed <- 42
new_sbm_s4(nodes = nodes,
           edges = edges,
           node_types = node_types,
           random_seed = 42)

#> SBM S4 class with 7 nodes of types {a, b} and 6 edges.
#> There is no initialized block structure.