Simulate stochastic block model of given number of groups

Simulates a simple block structured network with a desired number of constant sized groups. The probability of connections between given groups will be drawn from a passed function that must return a numeric value that will be used as the propensity parameter for drawing edge counts from a node pair. The default values will draw propensity values for bernouli edges with an average of a 50% chance of connection.

sim_basic_block_network(
  n_groups = 2,
  n_nodes_per_group = 5,
  prob_of_connections_dist = function(n) rbeta(n = n, shape1 = 2, shape2 = 2),
  edge_dist = purrr::rbernoulli,
  allow_self_connections = FALSE,
  keep_connection_counts = FALSE,
  return_connection_propensities = FALSE
)

Arguments

n_groups
How many groups to simulate

n_nodes_per_group
How many nodes in each group

prob_of_connections_dist
Function that takes a single size argumenet and returns a vector of connection propensities of specified size. Default is drawing Bernouli probabilities from a symmetric beta distribution centered at 0.5

edge_dist
A distribution function that has two inputs: the first is number of samples to draw and the second the propensity value from connection_propensities. For instance, the default value is rpois which takes two arguments: n and lambda. In this scenario the propensity of connections between two nodes is provided to lambda and thus is the average number of connections for each pair of nodes between two given groups.

allow_self_connections
Should nodes be allowed to have connections to themselves?

keep_connection_counts
Should the connection counts stay on returned edges? If edges distribution is a binary yes or no then you will likely want to set this to TRUE.

return_connection_propensities
If set to TRUE the returned list will also include the simulated connection propensities dataframe. This can be used for recreating draws using sim_sbm_network().

Value

A list with a nodes dataframe (containing a node's id and group membership), an edges dataframe (containing from and to nodes along with the total number of connections as drawn from edge_dist), and if return_connection_propensities == TRUE: connection_propensities dataframe that shows the randomly drawn connection propensities between groups.

See also

Examples

sim_basic_block_network(n_groups = 4, n_nodes_per_group = 10)

#> $nodes
#> # A tibble: 40 x 2
#>    id    group
#>    <chr> <chr>
#>  1 g1_1  g1   
#>  2 g1_2  g1   
#>  3 g1_3  g1   
#>  4 g1_4  g1   
#>  5 g1_5  g1   
#>  6 g1_6  g1   
#>  7 g1_7  g1   
#>  8 g1_8  g1   
#>  9 g1_9  g1   
#> 10 g1_10 g1   
#> # … with 30 more rows
#> 
#> $edges
#> # A tibble: 366 x 2
#>    from  to   
#>    <chr> <chr>
#>  1 g1_1  g1_2 
#>  2 g1_1  g1_3 
#>  3 g1_1  g1_7 
#>  4 g1_1  g1_8 
#>  5 g1_1  g1_9 
#>  6 g1_1  g1_10
#>  7 g1_1  g2_1 
#>  8 g1_1  g2_3 
#>  9 g1_1  g2_4 
#> 10 g1_1  g2_5 
#> # … with 356 more rows
#>