Plot network of nodes connected by pairwise block propensity

Takes the results of mcmc_sweep() with track_pairs = TRUE and plots the data as a network where two nodes have an edge between them if their pairwise propensity to be in the same block is above some threshold.

visualize_propensity_network(sbm, proportion_threshold = 0.2)

Arguments

proportion_threshold
Threshold of pairwise propensity to consider two nodes linked. Choose carefully!

Value

An interactive network plot with force layout.

See also

Examples

set.seed(42)
# Simulate network data and initialize model with it
net <- sim_basic_block_network(n_blocks = 3, n_nodes_per_block = 30) %>%
  collapse_blocks(desired_num_blocks = 1, sigma = 1.1) %>%
  choose_best_collapse_state(verbose = TRUE) %>%
  mcmc_sweep(num_sweeps = 100, eps = 0.4, track_pairs = TRUE)

#> Choosing collapse with 3 blocks and an entropy of 3044.1742.

# Plot connection propensity network
visualize_propensity_network(net, proportion_threshold = 0.4)