Run a single MCMC sweep over nodes

Runs a single Metropolis Hastings MCMC sweep across all nodes at specified level (For algorithm details see Piexoto, 2018). Each node is given a chance to move blocks or stay in current block and all nodes are processed in random order. Takes the level that the sweep should take place on (int) and if new blocks blocks can be proposed and empty blocks removed (boolean).

mcmc_sweep(
  sbm,
  num_sweeps = 1,
  variable_num_blocks = TRUE,
  eps = 0.1,
  track_pairs = FALSE,
  level = 0,
  verbose = FALSE
)

Arguments

sbm
SBM model object as created by create_sbm().

num_sweeps
Number of times all nodes are passed through for move proposals.

variable_num_blocks
Should the model allow new blocks to be created or empty blocks removed while sweeping or should number of blocks remain constant?

eps
Controls randomness of move proposals. Effects both the block merging and mcmc sweeps.

track_pairs
Return a dataframe with all pairs of nodes along with the number of sweeps they shared the same group?

level
Level of nodes who's blocks will have their block membership run through MCMC proposal-accept routine.

verbose
If set to TRUE then each proposed move for all sweeps will have information given on entropy delta, probability of moving, and if the move were accepted printed to the console.

Value

List with two dataframes. The first telling for all sweeps everytime a node was moved and what group it was moved to. The second telling for each sweep the entropy delta and total number of nodes that were moved to new groups in that sweep. If track_pairs = TRUE, then an additional pairing_counts dataframe is added to output.

Examples

set.seed(42)

# Start with a random network and assign randomly to 4 blocks
n_blocks <- 3
my_sbm <- create_sbm(sim_basic_block_network(n_blocks = n_blocks, n_nodes_per_block = 15)) %>%
  initialize_blocks(num_blocks = n_blocks)

# Calculate entropy with random blocks
get_entropy(my_sbm)

#> [1] 752.6645

# Run 4 MCMC sweeps
sweep_results <- my_sbm %>% mcmc_sweep(num_sweeps = 4, variable_num_blocks = FALSE)

# Look at the per-sweep level information
sweep_results$sweep_info

#>   entropy_delta num_nodes_moved
#> 1    -2.9956111              22
#> 2     0.2155941              15
#> 3     1.3341808              18
#> 4     0.4081434              21

# Get idea of node-pair similarity by looking at how often every pair of nodes is connected over sweeps
my_sbm %>% mcmc_sweep(num_sweeps = 4, track_pairs = TRUE)

#> $nodes_moved
#>  [1] "g2_13" "g1_5"  "g3_8"  "g3_13" "g2_5"  "g3_10" "g1_9"  "g2_11" "g1_11"
#> [10] "g1_13" "g3_9"  "g1_1"  "g3_4"  "g1_14" "g2_3"  "g3_3"  "g1_2"  "g2_15"
#> [19] "g3_2"  "g2_6"  "g3_12" "g1_10" "g2_4"  "g1_3"  "g1_3"  "g1_6"  "g1_7" 
#> [28] "g2_4"  "g2_7"  "g3_11" "g2_12" "g2_5"  "g1_9"  "g2_13" "g3_15" "g1_15"
#> [37] "g3_12" "g2_3"  "g3_5"  "g1_12" "g2_2"  "g1_10" "g3_13" "g2_15" "g1_13"
#> [46] "g1_1"  "g1_2"  "g3_13" "g1_3"  "g2_14" "g1_8"  "g1_9"  "g2_1"  "g1_7" 
#> [55] "g1_4"  "g3_8"  "g3_9"  "g1_6"  "g1_5"  "g1_10" "g2_9"  "g2_7"  "g3_14"
#> [64] "g3_5"  "g1_14" "g3_7"  "g2_5"  "g1_1"  "g3_3"  "g3_8"  "g3_13" "g1_11"
#> [73] "g3_3"  "g1_4"  "g3_6"  "g3_15" "g1_1"  "g3_1"  "g3_10" "g2_14" "g3_9" 
#> [82] "g1_10" "g1_6"  "g1_2"  "g2_6"  "g1_13" "g1_5"  "g1_14" "g3_5"  "g2_12"
#> [91] "g1_12" "g1_7"  "g1_3"  "g2_11" "g2_8"  "g2_2" 
#> 
#> $sweep_info
#>   entropy_delta num_nodes_moved
#> 1    -0.4842890              24
#> 2     0.7937078              22
#> 3     0.2806334              23
#> 4    -1.0456080              27
#> 
#> $pairing_counts
#>     node_a node_b times_connected proportion_connected
#> 1     g3_9   g3_7               2                 0.50
#> 2     g3_8   g3_7               0                 0.00
#> 3     g3_9   g3_6               1                 0.25
#> 4     g3_7   g3_6               1                 0.25
#> 5     g3_9   g3_5               2                 0.50
#> 6     g3_8   g3_5               2                 0.50
#> 7     g3_7   g3_5               2                 0.50
#> 8     g3_6   g3_5               1                 0.25
#> 9     g3_9   g3_4               4                 1.00
#> 10    g3_8   g3_4               2                 0.50
#> 11    g3_7   g3_4               2                 0.50
#> 12    g3_7   g3_3               1                 0.25
#> 13    g3_6   g3_3               3                 0.75
#> 14    g3_4   g3_3               3                 0.75
#> 15    g3_9   g3_2               1                 0.25
#> 16    g3_7   g3_2               0                 0.00
#> 17    g3_3   g3_2               4                 1.00
#> 18    g3_9  g3_15               1                 0.25
#> 19    g3_7  g3_15               2                 0.50
#> 20    g3_4  g3_15               2                 0.50
#> 21    g3_3  g3_15               2                 0.50
#> 22    g3_9  g3_14               2                 0.50
#> 23    g3_5  g3_14               2                 0.50
#> 24    g3_9  g3_13               2                 0.50
#> 25    g3_8  g3_13               2                 0.50
#> 26    g3_7  g3_13               2                 0.50
#> 27    g3_6  g3_13               2                 0.50
#> 28    g3_5  g3_13               2                 0.50
#> 29    g3_3  g3_13               2                 0.50
#> 30    g3_2  g3_13               1                 0.25
#> 31   g3_15  g3_13               2                 0.50
#> 32    g3_9  g3_12               0                 0.00
#> 33    g3_2  g3_12               3                 0.75
#> 34   g3_15  g3_12               1                 0.25
#> 35   g3_14  g3_12               2                 0.50
#> 36   g3_13  g3_12               2                 0.50
#> 37    g3_8  g3_11               2                 0.50
#> 38    g3_7  g3_11               0                 0.00
#> 39    g3_6  g3_11               1                 0.25
#> 40    g3_5  g3_11               2                 0.50
#> 41   g3_12  g3_11               3                 0.75
#> 42    g3_9  g3_10               2                 0.50
#> 43    g3_8  g3_10               3                 0.75
#> 44    g3_7  g3_10               3                 0.75
#> 45    g3_5  g3_10               2                 0.50
#> 46    g3_4  g3_10               3                 0.75
#> 47    g3_3  g3_10               1                 0.25
#> 48   g3_12  g3_10               1                 0.25
#> 49    g3_8   g3_1               1                 0.25
#> 50    g3_5   g3_1               1                 0.25
#> 51    g3_4   g3_1               0                 0.00
#> 52    g3_3   g3_1               0                 0.00
#> 53    g3_2   g3_1               3                 0.75
#> 54   g3_15   g3_1               0                 0.00
#> 55   g3_12   g3_1               1                 0.25
#> 56   g3_11   g3_1               2                 0.50
#> 57   g3_10   g3_1               1                 0.25
#> 58    g3_9   g2_9               2                 0.50
#> 59    g3_7   g2_9               2                 0.50
#> 60    g3_6   g2_9               1                 0.25
#> 61    g3_3   g2_9               1                 0.25
#> 62   g3_15   g2_9               2                 0.50
#> 63   g3_14   g2_9               2                 0.50
#> 64   g3_12   g2_9               2                 0.50
#> 65    g3_1   g2_9               0                 0.00
#> 66    g3_9   g2_8               3                 0.75
#> 67    g3_7   g2_8               3                 0.75
#> 68    g3_5   g2_8               2                 0.50
#> 69    g3_4   g2_8               1                 0.25
#> 70   g3_15   g2_8               2                 0.50
#> 71   g3_14   g2_8               3                 0.75
#> 72   g3_13   g2_8               2                 0.50
#> 73   g3_12   g2_8               1                 0.25
#> 74   g3_11   g2_8               0                 0.00
#> 75   g3_10   g2_8               3                 0.75
#> 76    g2_9   g2_8               3                 0.75
#> 77    g3_7   g2_7               1                 0.25
#> 78    g3_5   g2_7               2                 0.50
#> 79    g3_3   g2_7               0                 0.00
#> 80    g3_2   g2_7               3                 0.75
#> 81   g3_15   g2_7               1                 0.25
#> 82   g3_14   g2_7               1                 0.25
#> 83   g3_12   g2_7               1                 0.25
#> 84   g3_10   g2_7               1                 0.25
#> 85    g2_9   g2_7               1                 0.25
#> 86    g3_9   g2_6               1                 0.25
#> 87    g3_7   g2_6               1                 0.25
#> 88    g3_6   g2_6               0                 0.00
#> 89    g3_5   g2_6               1                 0.25
#> 90    g3_3   g2_6               3                 0.75
#> 91    g3_2   g2_6               1                 0.25
#> 92   g3_15   g2_6               1                 0.25
#> 93   g3_14  g3_11               0                 0.00
#> 94   g3_10   g2_6               1                 0.25
#> 95    g3_2  g3_11               3                 0.75
#> 96    g3_1   g2_6               3                 0.75
#> 97    g2_9   g2_6               1                 0.25
#> 98    g3_9   g2_5               2                 0.50
#> 99    g3_7   g2_5               3                 0.75
#> 100   g3_3   g2_5               1                 0.25
#> 101   g3_2   g2_5               2                 0.50
#> 102  g3_15   g2_5               3                 0.75
#> 103  g3_14   g2_5               3                 0.75
#> 104  g3_13   g2_5               2                 0.50
#> 105   g3_4  g3_14               2                 0.50
#> 106  g3_12   g2_5               1                 0.25
#> 107  g3_11   g2_5               2                 0.50
#> 108  g3_10   g2_5               3                 0.75
#> 109   g2_9   g2_5               3                 0.75
#> 110   g2_8   g2_5               1                 0.25
#> 111   g2_7   g2_5               1                 0.25
#> 112   g2_6   g2_5               1                 0.25
#> 113   g3_7   g2_4               3                 0.75
#> 114   g3_4   g2_4               1                 0.25
#> 115  g3_11   g2_4               3                 0.75
#> 116   g3_5  g1_10               2                 0.50
#> 117  g3_10   g2_4               2                 0.50
#> 118   g3_1   g2_4               0                 0.00
#> 119   g3_9   g2_3               3                 0.75
#> 120   g3_8   g1_7               3                 0.75
#> 121   g2_9   g2_4               3                 0.75
#> 122   g2_8   g2_4               2                 0.50
#> 123   g2_7   g2_4               0                 0.00
#> 124   g2_5   g2_4               1                 0.25
#> 125  g3_14   g2_1               2                 0.50
#> 126   g3_7   g2_3               0                 0.00
#> 127   g3_5   g2_3               2                 0.50
#> 128   g2_5   g2_3               3                 0.75
#> 129   g2_4   g2_3               1                 0.25
#> 130   g3_4  g2_14               0                 0.00
#> 131   g3_3   g2_2               1                 0.25
#> 132   g3_2   g2_2               1                 0.25
#> 133  g3_15   g2_2               2                 0.50
#> 134   g2_9   g2_2               2                 0.50
#> 135   g2_8   g2_2               2                 0.50
#> 136   g3_2   g1_9               4                 1.00
#> 137   g3_8   g1_5               3                 0.75
#> 138   g2_7   g2_2               3                 0.75
#> 139   g3_9   g2_2               2                 0.50
#> 140   g2_4   g2_2               1                 0.25
#> 141   g2_3  g1_15               3                 0.75
#> 142   g3_9  g2_15               2                 0.50
#> 143   g1_4  g1_15               1                 0.25
#> 144   g3_7  g2_15               1                 0.25
#> 145  g2_13   g2_1               1                 0.25
#> 146   g3_5  g2_15               2                 0.50
#> 147   g3_8   g3_6               3                 0.75
#> 148   g3_4  g2_15               4                 1.00
#> 149   g3_3  g2_15               4                 1.00
#> 150  g3_15  g2_15               3                 0.75
#> 151  g3_12  g2_15               0                 0.00
#> 152   g2_4   g2_1               2                 0.50
#> 153  g3_10  g2_15               3                 0.75
#> 154   g3_1  g2_15               2                 0.50
#> 155  g2_14  g1_12               1                 0.25
#> 156   g2_6  g2_15               2                 0.50
#> 157  g3_15  g1_15               2                 0.50
#> 158   g3_7  g2_14               2                 0.50
#> 159   g3_3  g2_14               1                 0.25
#> 160   g3_2  g2_14               3                 0.75
#> 161  g3_13  g2_14               2                 0.50
#> 162  g3_11  g2_14               2                 0.50
#> 163   g2_5  g2_14               1                 0.25
#> 164   g2_2  g2_14               2                 0.50
#> 165   g3_9  g2_13               2                 0.50
#> 166   g3_2  g2_13               3                 0.75
#> 167   g3_5   g1_3               2                 0.50
#> 168  g3_13  g2_13               1                 0.25
#> 169   g1_6   g1_2               2                 0.50
#> 170  g2_15  g2_11               2                 0.50
#> 171  g3_12  g2_13               3                 0.75
#> 172  g2_11  g1_11               3                 0.75
#> 173   g3_1  g2_13               2                 0.50
#> 174   g2_7  g2_13               3                 0.75
#> 175   g2_8   g2_6               1                 0.25
#> 176   g2_6  g1_10               2                 0.50
#> 177   g2_6  g2_13               3                 0.75
#> 178   g1_9  g1_15               1                 0.25
#> 179   g2_8  g2_14               1                 0.25
#> 180  g2_12   g1_1               1                 0.25
#> 181   g2_5  g2_13               0                 0.00
#> 182  g1_15  g1_11               0                 0.00
#> 183   g2_4  g2_13               0                 0.00
#> 184   g2_2  g2_13               2                 0.50
#> 185  g2_15  g2_13               3                 0.75
#> 186   g3_5  g3_15               2                 0.50
#> 187   g3_8  g2_12               1                 0.25
#> 188  g2_10   g1_4               1                 0.25
#> 189   g3_4  g2_12               1                 0.25
#> 190  g3_15  g2_12               2                 0.50
#> 191   g2_4  g2_15               3                 0.75
#> 192   g3_9  g2_12               2                 0.50
#> 193  g3_14  g2_12               2                 0.50
#> 194   g3_3   g2_4               2                 0.50
#> 195   g2_2   g2_1               2                 0.50
#> 196  g3_10  g2_12               2                 0.50
#> 197   g3_1  g2_12               2                 0.50
#> 198   g2_9  g2_12               2                 0.50
#> 199  g2_11  g1_10               3                 0.75
#> 200   g2_6   g2_3               2                 0.50
#> 201  g1_14  g1_12               1                 0.25
#> 202   g2_8  g2_12               2                 0.50
#> 203  g3_13   g2_3               3                 0.75
#> 204   g2_6  g2_12               2                 0.50
#> 205   g2_5  g2_12               3                 0.75
#> 206   g2_4  g2_12               1                 0.25
#> 207   g2_2  g2_15               3                 0.75
#> 208   g2_2  g2_12               2                 0.50
#> 209   g3_5   g3_2               1                 0.25
#> 210   g1_9   g1_6               2                 0.50
#> 211   g3_6  g2_14               1                 0.25
#> 212  g2_15  g2_12               3                 0.75
#> 213  g2_14  g2_12               2                 0.50
#> 214   g3_7  g2_11               0                 0.00
#> 215   g3_6   g3_2               0                 0.00
#> 216   g3_5  g2_11               1                 0.25
#> 217   g3_3  g2_11               4                 1.00
#> 218   g3_7  g1_11               0                 0.00
#> 219   g3_4  g2_13               0                 0.00
#> 220  g3_15  g2_11               0                 0.00
#> 221   g3_1  g2_11               3                 0.75
#> 222   g2_2  g1_15               1                 0.25
#> 223   g2_9  g2_11               0                 0.00
#> 224   g2_7  g2_11               2                 0.50
#> 225  g3_13   g3_1               1                 0.25
#> 226   g2_6  g2_11               3                 0.75
#> 227   g3_6   g3_1               1                 0.25
#> 228   g2_5  g2_11               3                 0.75
#> 229  g2_10  g1_15               1                 0.25
#> 230   g2_4  g2_11               0                 0.00
#> 231  g2_15  g1_14               1                 0.25
#> 232   g3_5  g2_10               1                 0.25
#> 233  g3_13  g2_10               1                 0.25
#> 234   g3_1   g1_5               1                 0.25
#> 235   g2_9  g2_10               2                 0.50
#> 236  g2_12   g1_7               2                 0.50
#> 237   g2_6  g2_14               3                 0.75
#> 238   g2_6  g2_10               0                 0.00
#> 239   g2_4  g2_10               1                 0.25
#> 240  g2_12   g1_5               1                 0.25
#> 241  g2_14  g2_10               1                 0.25
#> 242   g1_2  g1_11               3                 0.75
#> 243  g2_13  g2_10               3                 0.75
#> 244   g3_7   g3_1               0                 0.00
#> 245   g3_6   g1_6               1                 0.25
#> 246  g2_12  g2_10               2                 0.50
#> 247   g2_5   g1_1               3                 0.75
#> 248   g3_9   g2_1               0                 0.00
#> 249   g3_1   g2_8               1                 0.25
#> 250   g3_1  g2_14               3                 0.75
#>  [ reached 'max' / getOption("max.print") -- omitted 740 rows ]
#>