Run a single MCMC sweep over nodes

Runs a single Metropolis Hastings MCMC sweep across all nodes at specified level (For algorithm details see Piexoto, 2018). Each node is given a chance to move groups or stay in current group and all nodes are processed in random order. Takes the level that the sweep should take place on (int) and if new groups groups can be proposed and empty groups removed (boolean).

mcmc_sweep(sbm, level = 0, variable_num_groups = TRUE, beta = 1.5)

Arguments

sbm
SBM model object as created by create_sbm().

level
Level of nodes who's groups will have their group membership run through MCMC proposal-accept routine.

variable_num_groups
Should the model allow new groups to be created or empty groups removed while sweeping or should number of groups remain constant?

beta
Inverse temperature parameter for determining move acceptance probability.

Value

List with array of all nodes that were moved in the sweep (nodes_moved) and the sweeps impact on the model's entropy (entropy_delta).

Examples

set.seed(42)

# Start with a random network and assign randomly to 4 groups
n_groups <- 3
my_sbm <- create_sbm(sim_basic_block_network(n_groups = n_groups, n_nodes_per_group = 15)) %>%
  initialize_groups(num_groups = n_groups)

# Calculate entropy with random groups
compute_entropy(my_sbm)

#> [1] 752.6645

# Run 4 MCMC sweeps
for(i in 1:4) my_sbm %>% mcmc_sweep(variable_num_groups = FALSE)

# Calculate entropy after sweeps
compute_entropy(my_sbm)

#> [1] 683.4771